msm: mdss: allow re-queue of writeback buffers
When a buffer has been properly processed, user space may choose to use
sync fence to acknowledge when buffer is complete. In such cases dequeue
is not not necessary, allow this use case by allowing client to re-queue
of writeback buffers as long as the buffer has been properly processed
already.

Change-Id: Ibac1eb9710477734e1ee2daf8136085d4439b0d4
Signed-off-by: Adrian Salido-Moreno <adrianm@codeaurora.org>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>