Update patch set 2
Patch Set 2: Code-Review-1

I would say that it is generally a good idea to separate the defconfig commit from the change. It is easier to manage. Some additions to the kernel isolate an option that is better to be excluded. The driver works really well on Onyx. I would assume the same on Hammerhead as well. I am just waiting for review. Look at the interactive tunings I made as well. They work really great with this driver

Patch-set: 2
Label: Code-Review=-1