Update patch set 1
Patch Set 1: Verified+1 Code-Review+1

I'm going to give my personal opinion on this. I added this to Onyx, because I personally thought the display was a little orange at times, and lowering R&G to 245 and slightly raising saturation helped a lot with that. I'm not sure if Hammerheads display really needs it as I don't own one, but it is surely something to think about, and if it is really bad, you might want to export these in the init IF it applies to all screens. I haven't talked with others to see what they have thought in the case of Onyx, but I would definitely do some research

Patch-set: 1
Reviewer: Gerrit User 1000166 <1000166@3162e3d8-d894-470e-abd9-724f7b0effad>
Label: Code-Review=+1
Label: Verified=+1